% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_words_all.R
\name{find_words_all}
\alias{find_words_all}
\title{Find Strings in All Columns}
\usage{
find_words_all(df, ...)
}
\arguments{
\item{df}{input data frame}

\item{...}{your strings for search. Must be lowercase}
}
\value{
Filtered Data Frame contains row that your strings looking for.
}
\description{
Find Strings in All Columns
}
