% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_words.R
\name{find_words}
\alias{find_words}
\title{Find One Strings in Specific Column}
\usage{
find_words(df, column, ...)
}
\arguments{
\item{df}{input data frame}

\item{column}{column that you search your script}

\item{...}{your strings for search. Must be lowercase}
}
\value{
Filtered Data Frame contains row that your strings looking for.
}
\description{
Find One Strings in Specific Column
}
\examples{

}
